Update grid alignment tool.ahk
Prevented lines bleeding into other monitors when changing the thickness near the edge of the screen. 

Made sure the tooltip stayed in a relatively intuitive spot when the lines were near the edges of the screen. There was one scenario I tested that resulted in the tooltip ending up on another monitor but only when clicking on the taskbar and the primary monitor was on top of another monitor. No idea why but I can't imagine this affecting anyone.

Simplified some code.